ByRole deserves it own file
